main.py
writing code for testing.
We could just substitue lfm_360k with the name of the file in config.ini to test. the data is read successfully but the usage of memory is unbelievably high. Isa could try this test for lfm_100k since she has the dataset locally. Note also that df is a dictionary so after print we try to find a way to load the file.