Dataset abstraction with preprocessing
Added base processor with reusable functionality for each dataset.
Added processors for the following datasets:
- LFM-1B
- LFM-2B
- LFM-360K
- ML-100K
- ML-25M

Updated Dataset class with the uniform interface that comes out at the other side of the data processing step. Each dataset gives easy access to (information of) data tables.
Added three helper functions to provide formatting of a generic dataframe that contains a 'user' and/or 'item' column.

Updated data registry to automatically preprocess a dataset when the configuration of the dataset is not yet available

Added some utility functions for loading/saving and reading/writing tables.
Minor updates to reflect the previous changes.

Added preprocessed ML-100K dataset to the tests folder.