Refactoring algorithm tests
Modified the predictor/recommender interface tests into smaller chunks of code.
Modified Elliot recommender to return RuntimeError's instead of Notimplemented.
Added test to check if the subset dataframe is sufficient to run algorithm tests on.
Added extra assertions to check if the header in a returned dataframe is correct.
Added extra assertion to check prediction batching for multiple users.